+++ src/Like/Observers/LikeObserver.php
@@ -16,10 +16,10 @@ namespace Cog\Laravel\Love\Like\Observers;
 use Cog\Contracts\Love\Like\Models\Like as LikeContract;
 use Cog\Contracts\Love\Likeable\Services\LikeableService as LikeableServiceContract;
 use Cog\Laravel\Love\Like\Enums\LikeType;
-use Cog\Laravel\Love\Likeable\Events\ModelWasDisliked;
-use Cog\Laravel\Love\Likeable\Events\ModelWasLiked;
-use Cog\Laravel\Love\Likeable\Events\ModelWasUndisliked;
-use Cog\Laravel\Love\Likeable\Events\ModelWasUnliked;
+use Cog\Laravel\Love\Likeable\Events\LikeableWasDisliked;
+use Cog\Laravel\Love\Likeable\Events\LikeableWasLiked;
+use Cog\Laravel\Love\Likeable\Events\LikeableWasUndisliked;
+use Cog\Laravel\Love\Likeable\Events\LikeableWasUnliked;
 
 /**
  * Class LikeObserver.
@@ -37,10 +37,10 @@ class LikeObserver
     public function created(LikeContract $like)
     {
         if ($like->type_id == LikeType::LIKE) {
-            event(new ModelWasLiked($like->likeable, $like->user_id));
+            event(new LikeableWasLiked($like->likeable, $like->user_id));
             app(LikeableServiceContract::class)->incrementLikesCount($like->likeable);
         } else {
-            event(new ModelWasDisliked($like->likeable, $like->user_id));
+            event(new LikeableWasDisliked($like->likeable, $like->user_id));
             app(LikeableServiceContract::class)->incrementDislikesCount($like->likeable);
         }
     }
@@ -54,10 +54,10 @@ class LikeObserver
     public function deleted(LikeContract $like)
     {
         if ($like->type_id == LikeType::LIKE) {
-            event(new ModelWasUnliked($like->likeable, $like->user_id));
+            event(new LikeableWasUnliked($like->likeable, $like->user_id));
             app(LikeableServiceContract::class)->decrementLikesCount($like->likeable);
         } else {
-            event(new ModelWasUndisliked($like->likeable, $like->user_id));
+            event(new LikeableWasUndisliked($like->likeable, $like->user_id));
             app(LikeableServiceContract::class)->decrementDislikesCount($like->likeable);
         }
     }